+++ webui/src/ImportExport/Import/Page.tsx
@@ -10,12 +10,11 @@ import { ModuleInstanceType } from '@companion-app/shared/Model/Instance.js'
 import { ButtonGridHeader, PageNumberPicker, type PageNumberOption } from '~/Buttons/ButtonGridHeader.js'
 import { ButtonGridZoomControl } from '~/Buttons/ButtonGridZoomControl.js'
 import {
-	ButtonGridIcon,
-	ButtonGridIconBase,
 	ButtonInfiniteGrid,
 	type ButtonInfiniteGridButtonProps,
 	type ButtonInfiniteGridRef,
 } from '~/Buttons/ButtonInfiniteGrid.js'
+import { ButtonGridIcon, ButtonGridIconBase } from '~/Buttons/GridButtonIcons.js'
 import { useGridZoom } from '~/Buttons/GridZoom.js'
 import { Button } from '~/Components/Button'
 import { Callout } from '~/Components/Callout'
